changed package.json and seed.js to enable seeding the database separately (for Heroku)
